IGNITE-24950 / rename interface
