+++ modules/network/src/main/java/org/apache/ignite/internal/network/recovery/RecoveryServerHandshakeManager.java
@@ -53,7 +53,7 @@ import org.apache.ignite.internal.network.recovery.message.HandshakeRejectionRea
 import org.apache.ignite.internal.network.recovery.message.HandshakeStartMessage;
 import org.apache.ignite.internal.network.recovery.message.HandshakeStartResponseMessage;
 import org.apache.ignite.internal.network.recovery.message.ProbeMessage;
-import org.apache.ignite.internal.version.ProductVersionSource;
+import org.apache.ignite.internal.version.IgniteProductVersionSource;
 import org.apache.ignite.network.ClusterNode;
 
 /**
@@ -101,7 +101,7 @@ public class RecoveryServerHandshakeManager implements HandshakeManager {
 
     private final BooleanSupplier stopping;
 
-    private final ProductVersionSource productVersionSource;
+    private final IgniteProductVersionSource productVersionSource;
 
     /** Recovery descriptor. */
     private RecoveryDescriptor recoveryDescriptor;
@@ -124,7 +124,7 @@ public class RecoveryServerHandshakeManager implements HandshakeManager {
             ClusterIdSupplier clusterIdSupplier,
             ChannelCreationListener channelCreationListener,
             BooleanSupplier stopping,
-            ProductVersionSource productVersionSource
+            IgniteProductVersionSource productVersionSource
     ) {
         this.localNode = localNode;
         this.messageFactory = messageFactory;
