+++ modules/network/src/test/java/org/apache/ignite/internal/network/DefaultMessagingServiceTest.java
@@ -81,7 +81,7 @@ import org.apache.ignite.internal.network.serialization.marshal.UserObjectMarsha
 import org.apache.ignite.internal.testframework.BaseIgniteAbstractTest;
 import org.apache.ignite.internal.thread.NamedThreadFactory;
 import org.apache.ignite.internal.util.IgniteUtils;
-import org.apache.ignite.internal.version.DefaultProductVersionSource;
+import org.apache.ignite.internal.version.DefaultIgniteProductVersionSource;
 import org.apache.ignite.internal.worker.CriticalWorkerRegistry;
 import org.apache.ignite.network.ClusterNode;
 import org.apache.ignite.network.NetworkAddress;
@@ -534,7 +534,7 @@ class DefaultMessagingServiceTest extends BaseIgniteAbstractTest {
                 clusterIdSupplier,
                 clientHandshakeManagerFactoryAdding(beforeHandshake, bootstrapFactory, staleIdDetector, clusterIdSupplier),
                 channelTypeRegistry,
-                new DefaultProductVersionSource()
+                new DefaultIgniteProductVersionSource()
         );
         connectionManager.start();
         connectionManager.setLocalNode(node);
@@ -568,7 +568,7 @@ class DefaultMessagingServiceTest extends BaseIgniteAbstractTest {
                         clusterIdSupplier,
                         channel -> {},
                         () -> false,
-                        new DefaultProductVersionSource()
+                        new DefaultIgniteProductVersionSource()
                 ) {
                     @Override
                     protected void finishHandshake() {
